Split constants into focused modules (#26)
