+++ widgets/panels/card_box_panel.py
@@ -3,8 +3,8 @@ from typing import Any
 
 import wx
 
-from utils.constants import DARK_ACCENT, DARK_ALT, LIGHT_TEXT
 from utils.mana_icon_factory import ManaIconFactory
+from utils.ui_constants import DARK_ACCENT, DARK_ALT, LIGHT_TEXT
 
 
 class CardBoxPanel(wx.Panel):
