+++ widgets/panels/deck_research_panel.py
@@ -4,7 +4,6 @@ from collections.abc import Callable
 
 import wx
 
-from utils.constants import DARK_PANEL
 from utils.stylize import (
     stylize_button,
     stylize_choice,
@@ -12,6 +11,7 @@ from utils.stylize import (
     stylize_listbox,
     stylize_textctrl,
 )
+from utils.ui_constants import DARK_PANEL
 
 
 class DeckResearchPanel(wx.Panel):
