+++ widgets/panels/sideboard_guide_panel.py
@@ -10,8 +10,8 @@ from collections.abc import Callable
 import wx
 import wx.dataview as dv
 
-from utils.constants import DARK_ALT, DARK_PANEL, LIGHT_TEXT, SUBDUED_TEXT
 from utils.stylize import stylize_button
+from utils.ui_constants import DARK_ALT, DARK_PANEL, LIGHT_TEXT, SUBDUED_TEXT
 
 
 class SideboardGuidePanel(wx.Panel):
